Added a simple map creator
- A new map can now be created in the Client
- The new map can be choosen in the lobby
- The new map will be sent to the Server (if selected)
- Save name is hardcoded and cannot be changed from the game

+++ Shooterman/Client/GUI/InfoOverlay/InfoOverlay.cpp
@@ -51,7 +51,7 @@ void InfoOverlay::handleInfoMessages() {
 
 bool InfoOverlay::render(std::shared_ptr<sf::RenderWindow> window, sf::Vector2f mousePosition) {
   handleInfoMessages();
-  if (mClock.getElapsedTime() < sf::seconds(mNumberOfSecToShowMsg)) {
+  if (mClock.getElapsedTime() < sf::seconds((float)mNumberOfSecToShowMsg)) {
     mGuiFrame->render(window);
   }
   return true;
